pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "findit-keyframe"
version = "0.1.0"
description = "Extract temporally-distributed, quality-aware keyframes from video shots for VLM consumption."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Findit-AI" }]
keywords = ["video", "keyframe", "scene", "vlm", "siglip", "qwen"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Multimedia :: Video",
  "Topic :: Scientific/Engineering :: Image Processing",
]

dependencies = [
  "av>=11.0",
  "opencv-python-headless>=4.9",
  "numpy>=1.26",
  "pillow>=10.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-benchmark>=4.0",
  "ruff>=0.5",
  "pyright>=1.1",
]

[project.urls]
Homepage = "https://github.com/Findit-AI/findit-keyframe"

[tool.hatch.build.targets.wheel]
packages = ["src/findit_keyframe"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
  "E",    # pycodestyle errors
  "F",    # pyflakes
  "W",    # pycodestyle warnings
  "I",    # isort
  "N",    # pep8-naming
  "UP",   # pyupgrade
  "B",    # flake8-bugbear
  "SIM",  # flake8-simplify
  "RUF",  # ruff-specific
  "PTH",  # prefer pathlib over os.path
  "RET",  # flake8-return
]
ignore = [
  "E501",  # line too long (handled by formatter)
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.pyright]
include = ["src"]
strict = ["src"]
pythonVersion = "3.11"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
